feat(web,sdf,dal): Implement function editing
- Now the user can either edit or add a new qualification function to a
schema
- For now we do not support qualifications for component only
- In the process of developing this we noticed a bug in the node's
qualification icon state propagation that will be fixed in a upcoming PR
- Fixed a few type errors in the frontend

The frontend implementation consists mostly of QualificationEditor,
which can be refactored and improved, we should have some kind of SiCode
as base. But the PR is too big to think about this now.

The backend logic implied in fixing a few visibility bugs as we have to
clone the qualification function to our tenancy/visibility to be able to
edit it.