feat(web,dal,sdf): add provider ids to round trip through frontend
This change ships relevant provider IDs from SDF via
`schematic/list_schema_varants` to the frontend so they can be provided
when calling `schematic/create_connection`.

Sockets and Providers (ImplicitInternal and Explicit) now have a similar
relationship between Nodes and Components. In other words, a Socket is
updated to be a generic connecting point on a Node that fronts for a
more concrete domain concept, in this case either an ExternalProvider or
an ExplicityInternalProvider. As a result, the constructors for these
providers now also construct a related socket and their function names
are updated to reflect this. Sockets now have an additional `kind` field
which is similar to the Node's `kind` field so that we will know what
the Socket is fronting for.

Future Work
-----------

* Sockets don't need a name field anymore, assuming that each
  corresponding underlying type has a name, so we should remove them.
  For the moment when a Provider-backed Socket is created, both names
  are set to the same value.
* The `Includes` Socket side of a Component was left alone in the schema
  builtins--I'm not 100% sure what we do here, but at the moment these
  Sockets have nothing behind them.
* In `app/web`, the `SchematicInputSocket` and `SchematicOutputSocket`
  types are tightly coupled to providers which we can fix in a way
  similar to the `Node`/`Component` relationship, that is with some
  either types that have different fields depending on `kind`.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>