fix(dal): relax uniqueness constraint across attribute prototype args
This change addresses an issue where a user tries to create 2
connections from the same source to 2 seperate targets of the same type.
For example:

```mermaid
graph TD
    A[kubernetes_namespace: si-1] --> B[kubernetes_deployment: si-2]
    A --> C[kubernetes_deployment: si-3]
```

Prior to this change a uniqueness violation was triggered by
`intra_component_argument` which was sent back to the frontend as the
source of the failed API call. This change drops the first constraint
and refines the second to include `internal_provider_id` and visibility
uniqueness.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>