feat(dal): Handle deep data structures generated during dependent value updating
The function that we're running to generate the updated AttributeValue
at any point in time while updating dependent values might generate a
deep data structure for a Prop. If that happens, we need to expand the
returned value out into AttributeValues for each child Prop, instead
of storing it all as a single AttributeValue for the containing
Array/Map/Object.