fix(web): Fix panels reactivity
Also fix socket alignment after recent UI changes

It currently ignores writes in attribute views if component is changed
before blur, with the current architecture there isn't really a nice way
of keeping the PropertyEditor component living long enough with the old
component id for it to trigger blur. The parents just swap it out.

Instead of swapping and then blurring we clear the local data and swap
to avoid the blur on the wrong component.

Changed websocket event to not be Replay as it was retriggering the
toast even without changes.