bug(dal): Look for ExternalProviders that have the same tail Component as the value that was just updated
When looking for which `ExternalProvider`s use the value from an
`InternalProvider` that was just updated, we need to make sure the
`AttributeValueArgument` has the same tail `ComponentId`, instead of
the same head `ComponentId`. The tail is the _source_ side of the
relationship, whereas the head is the _destination_side.