feat(dal): Populate K8s Deployment container name from Docker Image
We now send the full tree of AttributeValues to the ExternalProvider
on DockerImage, so that we can get both the domain Props, and the SI
Name Prop, when building the Containers array on a K8s Deployment.

This also makes sure that the name for the container only contains
letters/numbers/dashes, and that it ends in letters/numbers.

Fixes ENG-238.