feat(dal): Move AttributeValue::list_payload_for_read_context to be a DB function
The original query would potentially find more AttributeValues than
were appropriate for the given context.

The new DB function does a breadth-first walk of the AttributeValue
tree, selecting the most appropriate child AttributeValues for the
list of AttributeValues that it has found at each step. So, instead of
attempting to find the most appropriate AttributeValues in a single
step, it will only consider the child AttributeValues of
AttributeValues that it has already determined were appropriate to
look at.