bug(dal): Only create unset child AttributeValues for Objects; do not propagate key to child AttributePrototypes
Previously, we would create an unwanted child
AttributeValue/AttributePrototype if the current Prop was an
Array/Map/Object, but we only wanted to do this if the Prop was an
Object. The child AttributeValues/AttributePrototypes for Arrays/Maps
are handled by AttributeValue::insert_for_context adding a new element
to the Array/Map, not by having "placeholder" values to represent that
something _could_ be there.

Also, we were previously using the key for the current value as the
key for the child AttributePrototypes that were being created, when
they should not have had a key. This was resulting in the placeholder
prototypes having the same key as the Array/Map entry that was just
added.