feat(dal): Be more selective of when to create/use AttributePrototypes & AttributeValues
Now, we only create the "base" AttributeValues & AttributePrototypes
for the "unset" value, starting at the root Prop, and up until we
reach an Array or Map (including the Array/Map). This means that a
Prop will no longer be guaranteed to have an AttributeValue &
AttributePrototype, but does mean that Arrays/Maps will no longer have
unnecessary "unset" values/prototypes sitting around, causing issues
for the front-end, and that are never used.

We also ensure that when we are adding an element to an Array/Map,
that the Array/Map itself exists in the same AttributeContext as the
element that we are trying to add to it. This means that the IndexMap
on the ArrayMap will only contain entries for a single
AttributeContext, and not for an AttributeContext and every more
specific AttributeContext than that one.