chore(dal,sdf): Increase stack size for Tokio runtime itself
We were already increasing the stack size for any tasks/threads that
Tokio would spawn, but the size of the stack for the thread that Tokio
itself was running in was still the default size. Now we spawn a new
thread that is configured to have the same stack size that Tokio will
use for threads that it creates, and start the Tokio runtime in that
new thread.

This does mean that we have an "extra" parent thread sitting around
doing nothing other than waiting for Tokio's thread to end, but
threads are cheap enough that this shouldn't really matter.

Now that Tokio itself is running in a thread with a larger stack size,
we no longer need the tests to run with a higher-than-default
optimization level to reduce the size of Tokio's stack usage.