chore(dal,sdf,pinga): Convert to trait based API for background jobs
By going to a trait-based interface, we're able to simplify a lot of
the code that will end up dealing with jobs that have been enqueued to
be sent off to the background job processing system, as well as the
jobs that the background job processing system is picking up to work.

This does come at the expense of using `Box<dyn Trait>`, but now that
the skeleton is in place, this should be relatively minor to maintain.

Addresses ENG-317