feat(sdf,dal): initial report job of failures
Reports job execution failures to the database using the job's
access_builder and visibility. If they can't be extracted we will only
log the failure, as it means a structural problem that shouldn't reach
the user, even with universal write, etc. We need to think about
integrating with error reporting systems.

Also refactors quite a bit of the executor, although it's not there yet,
we need the producer refactor to be merged to conciliate the
implementations.

Co-Authored-By: Victor Bustamante <victor@systeminit.com>
Co-Authored-By: Paulo Cabral Sanz <paulo@systeminit.com>