feat(web): status bar basics
This PR refactors the layout for the WorkspaceView, adds a new component
for our "Sidebars", and adds several new components to create Status
Bars that expand into tabs when clicked.

There is more work to be done here - I think we're close to being able
to refactor one more time to make this layout be driven more from the
Page, so that creating our other main navigation layouts will be easier.
But the functionality is all here, and it points the way to a structure
that should be pretty easy to work with.

An example:

``html
<StatusBarTab :selected="selected">
  <template #icon><CheckCircleIcon class="text-[#22C55E]" /></template>
  <template #name>Qualifications</template>
  <template #summary>
    <StatusBarTabPill>
      Total: <span class="font-bold">&nbsp; 3</span>
    </StatusBarTabPill>
    <StatusBarTabPill class="bg-[#DCFCE7] text-[#22C55E] font-bold">
      <CheckCircleIcon class="text-[#22C55E] w-4" />
      <div class="pl-px">3</div>
    </StatusBarTabPill>
  </template>
</StatusBarTab>
``

Shows a custom status bar tab, with icons, a name, and a summary
comprised of rounded pills. Eventually, that block should become its own
organism, using the molecules for the StatusBarTab and the StatusBarTabPill atom.

Baby steps!

Signed-off-by: Adam Jacob <adam@systeminit.com>