bug(ci): check for required artifacts before running compose
We will now fail in CI if we try and run the docker compose step without
the required artifacts being present in the repository. If we fail at
this step, it's likely because some *other* part of CI has failed
(because we failed to build the required artifacts) - but this will stop
the CI runner from having arbitrary root owned directories created when
we try and bind-mount to them.

Love,
Adam, Victor and Nick