chore(ci): refactor ci
This is a bug PR, that refactors the CI process to be more flexible. The
goal is faster, more targeted CI runs, and more use of the "top level"
makefile.

Each component now has a `build//`, `test//` and `lint//` task which is
aware of the required dependencies. For example:

```sh
$ make build//bin/veritech
```

Will build veritech, but is aware of the fact that it *also* needs a
build of cyclone and lang-js first. These dependencies are specified by
putting custom make targets in the top level `Makefile`.

It refactors our `ci` run to trigger a script (mk/test-changed.sh) that
looks at the files that have changed in our component directories, and
issues a single call to `make` with the test target for each specified.
So if a change impacted `app/web`, `lib/sdf` and `lib/dal`, it would
wind up calling:

```sh
$ make test//app/web test//lib/sdf test//lib/dal
```

Make will then figure out a correct order for all the targets
(including dependent targets), and execute only the impacted tests.

It also introduces `sccache` to our build system, which will cache build
artifacts from `rustc` between invocations of the compiler, without
manipulating the default `target` directory.