chore(ci): fix pinga image, change mount syntax
This fixes up the pinga image so that it will run in docker - because it
hijacks sdf, rather than using the dal directly, it needs its
configuration mounted into the `sdf` directory rather than `pinga`.

We also convert the bind mounts to the "long" syntax, which will error
if the source doesn't exist (rather than creating a directory, as the
short form does). This should cause our job to actually fail
appropriately if pinga doesn't exist and we try to bind mount it.