chore(pinga): Gracefully shutdown on faktory quiet state
When faktory sends a worker the "quiet" state in response to a
heartbeat, it means that the worker should not pick up any new work,
and shut down after it has completed what it's already working on.

The previous version of this was assuming that the "quiet" state only
meant that the worker should not pick up any new work.