feat(pinga): Terminate in-flight jobs on faktory terminate state
If the faktory server sends us the "terminate" state, we are supposed
to kill any in-flight jobs, and exit as soon as possible.

We now spawn the job in a sub-task, so that we can check to see if it
has finished, while polling the faktory client's heartbeat state. If
we see that the heardbeat state has changed to "terminate", then we
abort the task we spawned for the job. If the task has finished before
the heartbeat state has changed to "terminate", then we stop polling,
so we can return the result of the task.