feat(pinga): Gracefully shutdown on SIGINT/SIGTERM
We now hand a channel off to the various execution tasks, so that they
can look for a notification from a signal handler that is watching for
SIGINT/SIGTERM. Within the individual tasks that are doing the
fetch/execute loop for jobs, we check the channel between each fetch,
so that we end the task if a SIGINT has been seen.

The task that handles sending heartbeats to the faktory server does
not watch this main channel directly, as it needs to stick around at
least as long as any of the fetch/execute tasks are still around. We
set up a separate channel for it, that we send a notification on after
the fetch/execute tasks have all exited. This prevents the faktory
server from hanging up on us while we're still executing jobs, and
will need to be able to ACK/FAIL them once the job has finished.

Co-authored-by: zack@systeminit.com