chore(dal): Create SchemaVariant.finalize() & ensure it is used everywhere
Previously, it was a little too easy to forget to call both
`SchemaVariant::create_default_prototypes_and_values()` and
`SchemaVariant::create_implicit_internal_providers()` when
"finalizing" a  `SchemaVariant`. This adds a
`schema_variant.finalize()` method that encapsulates calling both, so
that there is only a single method that needs to be called.

`SchemaVariant::create_default_prototypes_and_values()` is still a
`pub` method, because our helper method for the builtins to create a
prop with a default value does not work appropriately if the parent
prototypes & values for the Prop that will be created do not already
exist.

Anywhere that we were calling
`SchemaVariant::create_default_prototypes_and_values()` to "finalize"
a `SchemaVariant` now uses this new `schema_variant.finalize()`
method, instead.