feat(sdf,web): impl `ChangeSetPanel` button behaviors
This change adds behavior to the `ChangeSetPanel` buttons, namely the
`Apply`, `Discard`, and `Commit` buttons, as well as refines some of the
change set selector behavior.

It turns out that each of these buttons performs more than one activity
in the backend (i.e. through SDF and using the DAL code) so it made more
sense for the frontend to call some new API endpoints to perform these
higher level actions.

So, what each button do high level and a bit more in detail?

Apply
=====

In our current user experience (I'll drop any mention of "old" or
"prior" behavior after this so if you're confused it's because I'm
talking new-and-current world!), and in the Compose view, our user is
working with an open change set and an started edit session. When
`Apply` is clicked:

- the currently opened edit session is *saved* (the edit session version
  of commiting/merging/saving down to a change set)
- then the associated and currently opened change set is *applied* (the
  change set version of commiting/merging/saving down to the main head)

Now we're left in a state without a started edit session or an open
change set, so we're going to redirect the user to the View view (i.e
the eyeball).

Commit
======

The action to take here is save all work from the last commit into the
change set and continue editing, so when `Commit` is clicked:

- the current edit session is *saved* (the edit session version of
  commiting/merging/saving down to a change set)
- then a new edit session is started from the still-current change set

As we're still in a good state (i.e. we have a started edit session and
an open change set) we're all done.

Note: You may have noted that we're *saving* an edit session here, but
clicking a button called "Commit". This is one facet that we'd like to
test with users based on results of prior user testing. The equivalent
button in the old interfaces was "Save" but we noticed this caused a few
users to click this button after every change they made in the diagram
or attribute editor. So, maybe tweaking the button name/verb/action to
take helps lower this confusion? We shall see! That's the reason that we
left the underlying rxjs service names and SDF API routes in the
language of our internal DAL domain (i.e. you *save* and *cancel* edit
sessions) rather than renaming this all the way down. It might make
sense to update our domain language in the future but not worth the
effort if our users are still confused with a "Commit" button, right?

Discard
=======

The intent here is to throw away any changes made from the last commit
to now, so when `Discard` is clicked:

- the current opened edit session is *canceled* (the edit session
  version of aborting/deleting/cancelling all changes in the edit
  session)
- then a new edit session is started for the still-current change set

As we're still in a good state, that is a started edit session and an
open change set, we're all done.

As with the "Commit" button, we're trying a user-facing label of
"Discard" rather than "Cancel" to see if that will help lower confusion.

---

Updated Change Set Selection behavior
=====================================

Knowing that we always want a started edit session in Compose view,
leaves us with some interesting states to consider. There will be
several instances where we don't want our user to be in a state where
they think they can make changes but actually cannot (because they need
to have a change set selected, for example). Here's a rundown of how
we're handling some of these scenarios:

No Open Change Sets
-------------------

In this case we want to force our user to create a change set, so in
this case a "create" modal dialog is presented. In this case, hitting
`ESC` on the keyboard or clicking the `Cancel` button will redirect the
user to the View view.

No Currently Selected Change Set With Open Change Sets
------------------------------------------------------

This is slightly different, in that we could absolutely drop our user
into the Compose view, but we need them to tell us which change set to
use. In this case a "select" modal dialog is presented which uses the
exact same change set selector component (including sharing the
underlying Vue state).

Hitting the `ESC` key or clicking the `Cancel` button here will also
redirect the user to the View view.

Note that as we're using the same change set selector there is the
`- new -` option present which allows our user to create a new change
set rather than selecting an open one. In this case we swap out the
presented "select" dialog for the "create" dialog. The creation flow is
the same as before, however if our user hits the `ESC` key or clicks the
`Cancel` button, the "select" dialog is swapped back in and our user can
select an already-opened change set. This fallback happens because our
panel component still has a list of opened change sets and it knows that
no change set/edit session pair is selected so the right behavior is to
force our user to choose a change set. Neat, huh?

Navigating To Compose View With A Change Set And Edit Session Selected
----------------------------------------------------------------------

A bit a of lengthy title, but that's the state, folks. In this case we
can ensure the change set selector has the right change set selected and
carry as normal.

Changing The Currently Selected Change Set
------------------------------------------

Remember that if we're in the Compose view without a model dialog
present then we should be able to edit/modify our system, and that this
is because we have a started edit session and open change set selected.
Both are necessary to mutate our
components/properties/nodes/connections/etc.

Now, what if our user decides to select another opened change set?

There is a chance that just before we switch them over, there are
changes present in the currently started edit session. So what should we
do with these changes (that is, abandon them, save them, throw up a
dialog asking our user what to do, etc)?

This code change (I mean Git commit here, not SI's commit/change
set/etc, lol that's tricky) will automatically on changing chanage sets:

- *save* the currently selected edit session
- update the currently selected change set
- *start* a new edit session for the newly selected change set

These activities like the button behaviors above are done in one SDF/API
call to ensure everything happens in lock-step.

---

Work Still Outstanding
======================

Two primary items of work remain to further polish this work:

- Filling in the dialog language, explaining about change sets and why
  we must create/select a change set where appropriate
- Fixing the overflow/shifting behavior of the change set selector in
  the "select" modal dialog. I bet I can get some help here, but found
  it rather late in the process.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>