bug(web): attributes no longer reactive
This PR fixes an issue where we were no longer issuing `WsEvent` from
Pinga when we vinished updating the dependent values of a component. We
now *do* send that when we update a value, which causes the UI to
automatically show the correct new data.

It also adds `rx-spy` as a dependency, which makes debugging RxJS observables
much easier.

We also removed some of the code that allowed for "stop" functionality
in Faktory to work, as it made each job take longer to execute. We also
lowered the sleep amount in the core faktory calling loop.

Signed-off-by: Adam Jacob <adam@systeminit.com>