fix(ci): ensure promote-image.sh fails when appropriate
This change fixes the `promote-image.sh` which should fail when Docker
pulls in an image with a specfic tag which doesn't exist.

Additionally, as every instance of `promote-image.sh` was the same, a
single version is extracted into `mk/promote-image.sh` and symlinked
back into their expected locations under a component's `script/`
directory.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>