chore(cyclone): refactor lib/cyclone into parts for to help compilation
This change splits the `lib/cyclone` component into 3 smaller, single
purpose crates. The refactoring here eliminates the feature sets present
in the former `lib/cyclone` component which housed 3 distinct groups of
code: 1) the Cyclone server, 2) the Cyclone client, 3) the common
request and response types for both the server and client.

With this change, there are 3 crates in the Cargo workspace:

- `lib/cyclone-server`: the Cyclone server-as-a-library code
- `lib/cyclone-client`: the Cyclone client-as-a-library code
- `lib/cyclone-core`: the common request/response types for the Cyclone
  server and client implementations

Each crate has no feature sets and so is compiled only once, further
diminishing re-compilation work we see in development (note that the
over rebuilding isn't totally eliminated, but would be further reduced
with a task to split Veritech in a similar way).

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>