fix(dal): Ensure selected AttributePrototypeArguments point to correct Component
In the case where we're trying to build up the arguments for the
`AttributePrototype` of an `InternalProvder` that is fetching data
from `ExternalProviders`, we need to make sure that the
`AttributePrototypeArgument`s have a head_component_id that matches
the `ComponentId` of our `AttributeWriteContext` from the
`AttributeValue` that we are trying to update.

If we don't ensure that the `AttributePrototypeArgument`s that we are
collecting are "pointing at" the correct `ComponentId`, then we end up
collecting _all_ `AttributePrototypeArgument` that are pointing at the
same `InternalProvider` for any `Component` of the same
`SchemaVariant`, which is definitely not what we want here.