fix(dal): trigger code generation in appropriate place
This is a minimal fix which restores triggering code generation on
component creation and attribute value updates. The prior implementation
had a block of code which wasn't being run on internal providers (and
it assumed that it was). Instead, we do the prior logic in
InternalProvider::implicit_emit which does at the appropriate time.

Co-authored-by: Jacob Helwig <jacob@systeminit.com>
Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>
Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>