feat(web,dal,sdf): resolve and run workflows e2e
Implement WorkflowRunner tab, allows you to pick a global workflow or a workflow for a specific component. Resolves the workflow, showing the execution plan. Allows you to run the workflow, consuming the output.

Currently everything is sync, no jobs get scheduled so resolve although cacheable waits on the HTTP request, running also waits on the HTTP request. We need a better way of informing the user the command they ran finished, in a appropriate interface to avoid mixing output runs in the current output sidebar.

Add sdf routes to list, resolve and run workflows.

Implement WorkflowPrototype that ties a context to a func. Implement WorkflowResolver that handles the workflow resolution (dry run), and caches it if the arguments stay the same.

Create builtins WorkflowPrototypes, for the workspace and kubernetes_deployments.

Workflow run output is sorted by execution start, not timestamp (as we don't have the timestamp granularity to properly sort it, but we could use index + timestamp eventually).