bug(dal): Fix tenancy & visibility checking in attribute_value_payload_for_root_stored_procedure
There were a couple of places where we were checking the wrong table's
visibility due to a copy/paste error. We're now using the
`in_tenancy_and_visible_v1` function to more concisely check both the
tenancy, and visibility at the same time, and checking against the
correct table in all places where we were checking just the visibility
before.

This also gets rid of a bunch of extraneous whitespace before a couple
of the queries, and removes the unnecessary usage of `EXECUTE` that
caused us to have to build up the query as a string with explicit bind
arguments, instead of writing the query directly, and letting Postgres
automatically replace the query with variable usage with a prepared
statement that has bind params for the variables.