feat(dal): Add function to generate graph data of AttributeValues affected by the given AttributeValue
As part of trying to be less naive about propagating changes out from
a given AttributeValue, we now have a PL/PgSQL function that returns a
table where each row is an AttributeValueId, and AttributeValueId that
must be updated before that AttributeValueId can be updated. If an
AttributeValue requires multiple other AttributeValues be updated
first, then it will have multiple rows in the returned table (one for
each AttributeValue that must be updated first).

The graph data that we generate is not guaranteed to be acyclic.

There are still some refinements to be done around determining which
AttributeValues are affected by a given AttributeValue while
generating the graph. Currently we have the potential to find too
many, because we are not sufficiently checking to see if there is a
more specific AttributeValue than the one we are currently looking at
that would be more appropriate to use as the input for the affected
AttributeValue that we found.