chore(dal,sdf): Reduce number of Faktory jobs spawned for updating depenent AttributeValues
This moves a large portion of the logic into the DB for figuring out the
graph of affected AttributeValues, and generating ones that should now
exist.

Previously, we were creating a DependentValuesUpdate job per
AttributeValue that needs to be updated, but now we only generate a job
that points at the AttributeValue that has been directly updated, and
the job will internally walk the graph of updates that need to happen
because the starting AttributeValue has been updated.