chore: impl safe concurrent access to txns over `DalContext` clones
 ## Background

This is a rather large change, but comes from the initial desire to
write code like this:

```rust
async fn doing_work(ctx: DalContext) -> Result<()> {
    // Setup work...

    let mut work_tasks = JoinSet::new();

    loop {
        // Compute some work to do...

        for id in things_to_update {
            let task_ctx = ctx.clone();
            work_tasks
                .build_task()
                .spawn(do_the_work_task(task_ctx, id));
        }

        match work_tasks.join_next().await {
            Some(future_result) => {
                let result = future_result??;
                // Record the result...
            }
            None => break;
        }
    }
}
```

where importantly, we would like to clone the `DalContext` in such a way
that we can spawn a Tokio task with the cloned copy.

 ## Owning The Database Transaction

The challenge to solve boils down to this: how can we pass a database
transaction around in a way that it is owned and can be locked for
mutual access while at the same time being safe?

The Rust type which unlocked this puzzle (which lives in the `pg` module
of `si-data`) is this:

```rust
 #[ouroboros::self_referencing]
struct PgOwnedTransaction {
    conn: InstrumentedClient,
    #[borrows(mut conn)]
    #[covariant]
    txn: Option<InstrumentedTransaction<'this>>,
}
```

This type allows us to ship the transaction around with the underlying
client connection riding right alongside. The details may not be
important, but the takeaway here is that we're creating a self
referential struct which can be tricky to set up correctly. For more
background, this article about [Self Referential Structs in
Rust](https://arunanshub.hashnode.dev/self-referential-structs-in-rust)
is very useful.

Now that we can "own" a transaction, then we can wrap this in another
Rust type which can control concurrent access to the transaction. Enter
`PgSharedTransaction`:

```rust
 #[derive(Clone)]
pub struct PgSharedTransaction {
    inner: Arc<Mutex<PgOwnedTransaction>>,
    metadata: Arc<ConnectionMetadata>,
}
```

On this type we wrap all calls to the `InstrumentedTransaction<'_>` so
that consumers of this type don't need to understand the inner
implementation details and importantly don't have to manage unlocking a
mutex before making database query calls.

 ## Refactoring The Result Over Codebase

These 2 additional types unlocked the ability for us to drop any borrow
lifetimes on our `DalContext<'_, '_, '_>` type, which is now a
`DalContext`. It also implements the `Clone` trait, meaning that
`let new_ctx = ctx.clone();` is possible.

With the lifetimes stripped out, it also became possible for the
`DalContextBuilder` to start its own `Transactions` instance rather than
having to provide one when calling `builder.build(...)`. These changes
also made the `TransactionsStarter` not nearly as important, but also
changed what its purpose was: to hold on to the underlying connections
from which we could start a `Transactions`. So, `TransactionsStarter`
is renamed to `Connections`.

Now that `Transactions` instances are typically created and owned inside
a `DalContext`, this meant that our `commit()` method could be attached
directly on `DalContext`. There is one large item to point out though
for `.commit()` and `.rollback()`: these methods consume `self` and
therefore consume the `DalContext` instance. In practice in our
codebase this has always been fine and didn't present any refactoring
issues, but if needed, there are ways to still commit/rollback a
`DalContext` and hold on to underlying `Connections`.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>