refactor(dal,si-test-macros): rename `TransactionsStart` to `Connections`
This change finishes a large refactoring that made it clear what the
new/current pupose of `TransactionsStarter` is: to hold all related
service connections from which to start an associated `Transactions`. To
better reflect this, the `Connections` type has a renamed method of
`start_txns()` to reflect what is happening.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>