chore(dal): add Clippy lint to catch `commit_and_continue` usage
This change adds a specific Clippy lint warning when, as a developer, I
forget to remove a call to `commit_and_continue()` in an integration
test.

Ideally we shouldn't be committing tests with this call as a database
commit could infect the correctness of other tests. A warning will be
shown with these calls and in CI all warnings are flipped to errors.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>