bug(web,cyclone): fix small regressions
The work to switch to Pinia stores rather than RxJS missed sending the
eventChangeSetWritten$ to standardVisibility$ - the result was that the
UI was not refreshing appropriately in many cases when work was done on
the server side. There were previously a few other events that were also
included in standard visibility; I believe they are not being used in
the current version of the UI, but it's hard to say for certain without
a deeper analysis than time warrants.

There is also a small fix to disable sending the `--enable-confirmations`
flag to cyclone, as it appears to not be enabled in cyclone itself. This
makes qualifications run to completion again, rather than generate an
error.

Signed-off-by: Adam Jacob <adam@systeminit.com>