chore(veritech): refactor lib/veritech into parts to help compilation
This change splits the `lib/veritech` component into 3 smaller, single
purpose crates. The refactoring here eliminates the feature setts
present in the former `lib/veritech` compoonent which housed 3 distinct
groups of code: 1) the Veritech server, 2) the Veritech client, 3) the
common NATS subject logic for both the server and client.

With this change, there are 3 crates in the Cargo workspace:

- `lib/veritech-client`: the Veritech client-as-a-library code
- `lib/veritech-server`: the Veritech server-as-a-library code
- `lib/veritech-core`: the common NATS subject logic for the Veritech
  server and client implementations

Each crate has not feature sets and so is compiled only once, further
diminishing re-compilation work we see in development (note that the
over rebuilding isn't totally eliminated, but would be further reduced
with a task to split si-data in a similar way).

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>