chore(web): remove the open address link
It feels too magical - where it's placed makes it seem like a side
effect of the fix, which it isn't. And what happens when there is more
than one?

The fake fix screen feels *really* legit, so I'm removing this while we
figure out exactly how we would want to dress up the UI to expose things
like "links" to the user from a fix/resource.

Signed-off-by: Adam Jacob <adam@systeminit.com>