feat(PG): Auto-explain nested statements
auto_explain.log_nested_statements causes nested statements (statements
executed inside a function) to be considered for logging. When it is
off, only top-level query plans are logged. Given that we have a *lot*
of functions defined that execute multiple queries, it would be very
handy to get query plans for any of those queries that are slow.