bug(dal): Ensure builtin funcs are available to production
Previously, we were relying on the source tree being availble relative
to the binary, but this assumption is false in production. Instead,
we're now compiling the builtin func files into the SDF binary, and
going over the in-memory version of them to see which ones need to be
created during boot.

Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>