chore: extract app-specific code into `telemetry-application-rs` crate
This change splits the `lib/telemetry-rs` into 2 smaller, single purpose
crates. This refactoring is the final "de-feature-set"'ing work, finally
eliminating conditional feature sets within our component crate
ecosystem. Note that presently `lib/config-file` technically has feature
sets, however it was written to be generic and ultimately in a way that
could be extracted and made open source (we only use one permutation so
I'm discounting the other code paths in this crate for now).

With this change, there is 1 new crate in the Cargo workspace and 1
modified:

- `lib/telemetry-rs`: tracing/telemetry-related code that every crate
  component can and should use for logging/timing/tracing/etc. This
  allows us to maintain **one** dependency change for the tracing
  ecosystem rather than tracking it *n* times across *n* crates. Also,
  why the `-rs` suffix? Well, there may likely be a `lib/telemetry-ts`
  in our future, so why name grab :shrug:
- `lib-telemetry-application-rs`: tracing and opentelemetry setup only
  required for full binary programs (that is server, processes, and
  other CLI tooling). Ideally only components under `bin/` should depend
  on this crate, with all other crates depending on `lib/telemetry-rs`.
  Note that this library re-exports most of `telemetry-rs`'s types so a
  binary crate should not have to depend on both of these crates.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>