feat(dal): order by id desc to prefer newer values in standard model
We should avoid surfacing universal tenancy versions of objects if one
exists in a specific tenancy. We solve this naively by ensuring the
newest ids show up first in the standard model table views.

This may cause issues in the future if we want to remove something from
universal tenancy and make a new version of it, with an updated id.
Adding a `CASE WHEN is_universal IS TRUE THEN 0 ELSE 1 END ASC`
to the `ORDER BY` clause does not solve that problem, because
`DISTINCT ON (x)` requires us to order by `x` *first*, which means the
second ordering by `is_universal` is only relevant to us if there is
an id match. But the `only_one_live` constraints we have in place mean
that there cannot be two objects with the same id in the same
visibility, so the visibility ordering will already take care of
providing the correct object in those cases.